Crate iron_mountrouter

Structs

MethodPicker
MethodPicker provides an interface to apply different handlers for different http methods as middleware for the Iron framework.

NoMethod
NoMethod is error type returned to Iron if no handler was matched in MethodPicker

NoRoute
NoRoute is error type returned to Iron if no route was matched in Router

OriginalUrl
OriginalUrl serves as a key in request extensions to original url if router strips url for mounted handlers.

Router
Router provides an interface to apply different handlers to different routes as middleware for the Iron framework.

Type Definitions

Params
Params is stored in request extensions and gives access to captured dynamic parameters