Struct iron_mountrouter::OriginalUrl

pub struct OriginalUrl;

OriginalUrl serves as a key in request extensions to original url if router strips url for mounted handlers.

let ref original_url = req.extensions.get::<OriginalUrl>()